Add new components and remove unused code
