+++ ReactApp/src/components/staging/Staging.jsx
@@ -1,13 +1,11 @@
 import React from "react";
 
 import withStyles from "@mui/styles/withStyles";
-import { withRouter } from "react-router-dom";
 
 import { Link } from "react-router-dom";
 import Button from "@mui/material/Button";
-import RedirectToLogIn from "React-Automation-Studio/components/SystemComponents/RedirectToLogin.js";
 
-import TraditionalLayout from "React-Automation-Studio/components/UI/Layout/ComposedLayouts/TraditionalLayout.js";
+import TraditionalLayout from "React-Automation-Studio/components/UI/Layout/ComposedLayouts/TraditionalLayout";
 import Grid from "@mui/material/Grid";
 
 const styles = (theme) => ({
@@ -40,7 +38,6 @@ const Staging = (props) => {
       denseAppBar
       alignTitle="center"
     >
-      <RedirectToLogIn />
       <Grid
         container
         direction="row"
@@ -81,4 +78,4 @@ const Staging = (props) => {
   );
 };
 
-export default withRouter(withStyles(styles)(Staging));
+export default withStyles(styles)(Staging);
